added classes Autor, AuthorDao,AuthorResourse, modyfied BookResourse method with Response

+++ src/main/java/org/example/BookResourse.java
@@ -2,6 +2,7 @@ package org.example;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 import java.util.List;
 
 @Path("/books")
@@ -16,10 +17,18 @@ public class BookResourse {
     @GET
     @Path("/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Book getBookById(@PathParam("id") int id) {
+    public Response getBookById(@PathParam("id") int id) {
+        Book book = bookDao.getBookById(id);
 
-        return bookDao.getBookById(id);
+        if (book != null) {
+            return Response.ok().entity(book).build();
+        } else {
+            return Response.status(Response.Status.NOT_FOUND)
+                    .entity("Book with ID " + id + " not found")
+                    .build();
+        }
     }
+
     @POST
     @Path("/add")
     @Consumes(MediaType.APPLICATION_JSON)
@@ -29,14 +38,31 @@ public class BookResourse {
     @PUT
     @Path("/{id}")
     @Consumes(MediaType.APPLICATION_JSON)
-    public void updateBook(@PathParam("id") int id, Book book) {
-        book.setId(id);
-        bookDao.updateBook(book);
+    public Response updateBook(@PathParam("id") int id, Book book) {
+        Book existingBook = bookDao.getBookById(id);
+        if (existingBook != null) {
+            bookDao.updateBook(book);
+            return Response.noContent().build();
+        } else {
+            return Response.status(Response.Status.NOT_FOUND)
+                    .entity("Book with ID " + id + "not found")
+                    .build();
+        }
+
     }
     @DELETE
     @Path("/{id}")
-    public void deleteBook(@PathParam("id") int id) {
-        bookDao.removeBook(id);
+    public Response deleteBook(@PathParam("id") int id) {
+        Book existingBook = bookDao.getBookById(id);
+
+        if (existingBook != null) {
+            bookDao.removeBook(id);
+            return Response.noContent().build();
+        } else {
+            return Response.status(Response.Status.NOT_FOUND)
+                    .entity("Book with ID " + id + "not found")
+                    .build();
+        }
     }
 
 }
